requirements.txt
joblib>=0.17.0
matplotlib>=3.3.3
numpy>=1.19.4
openpyxl>=3.0.5
pandas>=1.1.4
pathlib2>=2.3.5
pysam>=0.16.0.1
scipy>=1.5.4
seaborn>=0.11.0
statsmodels>=0.12.1
xlrd>=1.2.0
pyyaml>=5.3.1